Revert "fmod is in. time to make it a proper system"
This reverts commit d58ff9c9d451cb4d13c4dd169016459701eb42fa.

+++ DystopiaEngine/DystopiaEngine/Source/Editor/SceneView.cpp
@@ -439,15 +439,15 @@ namespace Dystopia
 	void SceneView::DrawGizmoMul(const AutoArray<GameObject*>& _arr)
 	{
 		Math::Pt3D avgPos{ 0, 0, 0, 1.f };
-		size_t size = _arr.size();
+		unsigned int size = _arr.size();
 		for (auto& obj : _arr)
 		{
 			const auto pos = obj->GetComponent<Transform>()->GetGlobalPosition();
 			avgPos.x = avgPos.x + pos.x;
 			avgPos.y = avgPos.y + pos.y;
 		}
-		avgPos.x = avgPos.x / static_cast<float>(size);
-		avgPos.y = avgPos.y / static_cast<float>(size);
+		avgPos.x = avgPos.x / size;
+		avgPos.y = avgPos.y / size;
 		Math::Vec2 screenPos = GetWorldToScreen(avgPos);
 		float changeX = 0;
 		float changeY = 0;
